fix: remove type duplication
